rename HavenaskAnalyzer
