+++ elastic-fed/modules/havenask-engine/src/main/java/org/havenask/engine/HavenaskEnginePlugin.java
@@ -84,9 +84,9 @@ import org.havenask.env.ShardLock;
 import org.havenask.index.IndexModule;
 import org.havenask.index.IndexSettings;
 import org.havenask.index.analysis.AnalyzerProvider;
-import org.havenask.index.analysis.HavenaskJiebaAnalyzerProvider;
-import org.havenask.index.analysis.HavenaskSimpleAnalyzerProvider;
-import org.havenask.index.analysis.HavenaskSinglewsAnalyzerProvider;
+import org.havenask.engine.index.analysis.HavenaskJiebaAnalyzerProvider;
+import org.havenask.engine.index.analysis.HavenaskSimpleAnalyzerProvider;
+import org.havenask.engine.index.analysis.HavenaskSinglewsAnalyzerProvider;
 import org.havenask.index.engine.EngineFactory;
 import org.havenask.index.mapper.Mapper;
 import org.havenask.index.shard.IndexMappingProvider;
