+++ elastic-fed/modules/havenask-engine/src/test/java/org/havenask/engine/index/config/generator/SchemaGeneratorTests.java
@@ -38,9 +38,9 @@ import org.havenask.engine.index.config.Schema;
 import org.havenask.engine.index.mapper.DenseVectorFieldMapper;
 import org.havenask.index.IndexSettings;
 import org.havenask.index.analysis.AnalyzerScope;
-import org.havenask.index.analysis.HavenaskJiebaAnalyzer;
-import org.havenask.index.analysis.HavenaskSimpleAnalyzer;
-import org.havenask.index.analysis.HavenaskSinglewsAnalyzer;
+import org.havenask.engine.index.analysis.HavenaskJiebaAnalyzer;
+import org.havenask.engine.index.analysis.HavenaskSimpleAnalyzer;
+import org.havenask.engine.index.analysis.HavenaskSinglewsAnalyzer;
 import org.havenask.index.analysis.IndexAnalyzers;
 import org.havenask.index.analysis.NamedAnalyzer;
 import org.havenask.index.mapper.MapperService;
